Adaption to the CI build environment

+++ eagle/build.sh
@@ -3,8 +3,14 @@
 mkdir build
 cd build
 cmake ..
-make all
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 
+make all
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 
 cd ..
 rm -rf release
@@ -13,6 +19,4 @@ mkdir ./release/eagle
 cp ./build/src/eagle ./release/eagle/
 cp -r ./conf  ./release/eagle/
 
-#make clean
-#cd ..
-#rm -rf build
+exit 0
